Work around Firefox list marker positioning issue
Firefox displayed list markers too low when the first element in a list
element is a literal block. Removing the overflow property from the
pre element works around this. The overflow property seems to be
redundant, as the highlight div wrapping the pre also has overflow-x
set to display a scroll bar when necessary, and overflow-y doesn't seem to
have any use here.